% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\docType{data}
\name{demo_group}
\alias{demo_group}
\title{Demonstration Sample Groups for ssGSEA}
\format{
A factor with two levels (Control, Treatment).
}
\usage{
data(demo_group)
}
\description{
A factor assigning the 12 samples of \code{demo_expr} to Control and
Treatment groups (6 samples each).
}
\keyword{datasets}
