% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\docType{data}
\name{demo_rank_apop}
\alias{demo_rank_apop}
\title{Demonstration Ranked Gene List for Single-Group GSEA}
\format{
A decreasingly sorted named numeric vector.
}
\usage{
data(demo_rank_apop)
}
\description{
A decreasingly sorted named numeric vector. All genes from the
representative Apoptosis gene set are assigned large positive ranking
statistics, creating a clear enrichment signal.
}
\keyword{datasets}
