% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldeath_ssGSEA.R
\name{print.celldeath_ssgsea}
\alias{print.celldeath_ssgsea}
\title{Print Method for celldeath_ssgsea Objects}
\usage{
\method{print}{celldeath_ssgsea}(x, ...)
}
\arguments{
\item{x}{A \code{celldeath_ssgsea} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for celldeath_ssgsea Objects
}
